+++ app/signin/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-export default function Signin() {
+export default async function Signin() {
   return (
     <div>
       <div>page</div>
